Replace ecdsa with PyCA/cryptography
Dependency on ecdsa is a problem for all applications and users that
have strict crypto requirements and do not permit custom implementations
of cryptographic algorithms. The package even warns that it is
potentially insecure and does not protect against side channel attacks.
It's less of an issue for pubkeys, but still a compliance issue.

Cryptography can handle ecdsa SSH keys without help of ecdsa Python
package. The PR replaces the key verification code. I have also included
a (mostly untested) re-implementation of ecdsa.key.VerifyingKey.

Signed-off-by: Christian Heimes <christian@python.org>